sidebar created and has been imported to Dashboard.js
