--- src/CSS/Dashboard/Dashboard.css
@@ -1,24 +1,0 @@
-.dashboard__Sidebar {
-    height: 100%;
-    width: 10%;
-    position: fixed;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    overflow-x: hidden;
-    background-color: aqua;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between; /* Add this line */
-    
-}
-
-.top__Nav {
-    display: flex;
-    flex-direction: column;
-}
-
-.bottom__Nav {
-    display: flex;
-    flex-direction: column;
-}
